parameterize powerman path and simplify on/off
parameterize powerman path so that the agent can work on
a distro that installs it somewhere other than /usr/sbin.

simplify the on/off code to remove checks that are redundant.
the pacemaker fencing library verifies the results of an
on or off action by calling the fence agent again to query
status.